Complete RBAC test coverage for Shipyard Document Staging API
This commmit completes RBAC coverage for Shipyard Document Staging API,
for the APIs noted here [0]. For now, the goal is to meet the first use-case
of this plugin, which is to test RBAC for Shipyard. With this in mind, for RBAC
testing, we only care if a role has permission to an API in question. Therefore,
some of the more complex APIs are 'short circuit' tests - meaning only RBAC
permissions are checked and other expections are ignored.

[0] http://airship-shipyard.readthedocs.io/en/latest/API.html#document-staging-api